+++ aws/request/request_test.go
@@ -3,7 +3,6 @@ package request_test
 import (
 	"bytes"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"io"
 	"io/ioutil"
@@ -317,7 +316,7 @@ func TestRequestRecoverTimeoutWithNilBody(t *testing.T) {
 		{StatusCode: 200, Body: body(`{"data":"valid"}`)},
 	}
 	errors := []error{
-		errors.New("timeout"), nil,
+		errTimeout, nil,
 	}
 
 	s := awstesting.NewClient(aws.NewConfig().WithMaxRetries(10))
@@ -352,8 +351,9 @@ func TestRequestRecoverTimeoutWithNilResponse(t *testing.T) {
 		nil,
 		{StatusCode: 200, Body: body(`{"data":"valid"}`)},
 	}
+
 	errors := []error{
-		errors.New("timeout"),
+		errTimeout,
 		nil,
 	}
